ai_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
torch==2.3.1
transformers==4.44.2
sentencepiece==0.2.0
protobuf==5.27.3
google-genai==1.5.0
python-dotenv==1.0.1